+++ src/tool-prange.cpp
@@ -1048,7 +1048,7 @@ void Tool_prange::assignHorizontalPosition(vector<_VoiceInfo>& voiceInfo, int mi
 
 	if (hpos.size() > 2) {
 		for (int i=1; i<(int)hpos.size()-1; i++) {
-			int ii = hpos.size() - i - 1;
+			int ii = (int)hpos.size() - i - 1;
 			hpos[i] = (double)ii / (hpos.size()-1) * (maxval - minval) + minval;
 		}
 	}
@@ -1561,7 +1561,7 @@ int Tool_prange::getTopQuartile(vector<double>& midibins) {
 
 	double cumsum = 0.0;
 	int i;
-	for (i=midibins.size()-1; i>=0; i--) {
+	for (i=(int)midibins.size()-1; i>=0; i--) {
 		if (midibins[i] <= 0.0) {
 			continue;
 		}
@@ -1659,7 +1659,7 @@ int Tool_prange::getStaffBase7(int base7) {
 //
 
 int Tool_prange::getMaxDiatonicIndex(vector<vector<double>>& diatonic) {
-	for (int i=diatonic.size()-1; i>=0; i--) {
+	for (int i=(int)diatonic.size()-1; i>=0; i--) {
 		if (diatonic.at(i).at(0) != 0.0) {
 			return i;
 		}
